Reorganize keyboard code
