+++ src/device/keyboard/mod.rs
@@ -1,7 +1,13 @@
-use alloc::Vec;
 use alloc::string::{String, ToString};
-use device::{keyboard, ps2_controller_8042};
 use spin::Mutex;
+
+macro_rules! key_press {
+        ($x:expr) => (Some(KeyEvent::Pressed($x)))
+}
+
+macro_rules! key_release {
+        ($x:expr) => (Some(KeyEvent::Released($x)))
+}
 
 #[derive(Debug)]
 struct KeyPair {
@@ -36,7 +42,7 @@ pub enum Modifier {
 
 /// All of our supported keyboard modifiers.
 #[derive(Debug)]
-struct ModifierState {
+pub struct ModifierState {
     alt: KeyPair,
     caps_lock: bool,
     control: KeyPair,
@@ -63,7 +69,7 @@ impl ModifierState {
 
     /// Apply all of our modifiers to an ASCII character, and return a new
     /// ASCII character.
-    fn apply_to(&self, ascii: char) -> String {
+    pub fn apply_to(&self, ascii: char) -> String {
         if self.is_uppercase() {
             ascii.to_uppercase().collect()
         } else {
@@ -71,7 +77,7 @@ impl ModifierState {
         }
     }
 
-    fn update(&mut self, m: Modifier) {
+    pub fn update(&mut self, m: Modifier) {
         use self::Modifier::*;
 
         match m {
@@ -99,46 +105,6 @@ pub enum Key {
     LowerAscii(u8),
 }
 
-static STATE: Mutex<ModifierState> = Mutex::new(ModifierState::new());
+pub static STATE: Mutex<ModifierState> = Mutex::new(ModifierState::new());
 
-/// Get all bytes from keyboard and translate to key
-pub fn parse_key(scancode: u8) {
-    let byte_sequence: u64 = retrieve_bytes(scancode);
-    if let Some(key) = keyboard::get_key(byte_sequence) {
-        match key {
-            Key::Ascii(k) => print_char(k as char),
-            Key::Meta(modifier) => STATE.lock().update(modifier),
-            Key::LowerAscii(byte) => print_str(STATE.lock().apply_to(byte as char)),
-        }
-    }
-}
-
-/// Keep reading bytes until sequence is finished and combine bytes into an integer
-fn retrieve_bytes(scancode: u8) -> u64 {
-    let mut byte_sequence: Vec<u8> = vec![scancode];
-
-    // if byte is start of sequence, start reading bytes until end of sequence
-    // TODO: Design system that reads more than two bytes
-    if scancode == 0xE0 || scancode == 0xE1 {
-        let check_byte: u8 = ps2_controller_8042::key_read();
-        if let Some(byte) = keyboard::is_special_key(check_byte) {
-            byte_sequence.push(byte);
-        }
-    }
-
-    byte_sequence
-        .iter()
-        .rev()
-        .fold(0, |acc, &b| (acc << 1) + b as u64)
-}
-
-fn print_str(s: String) {
-    kprint!("{}", s);
-}
-
-fn print_char(byte: char) {
-    match byte {
-        '\n' | ' ' | '\t' => kprint!("{}", byte),
-        _ => (),
-    }
-}
+pub mod ps2;
