+++ src/device/mod.rs
@@ -1,6 +1,6 @@
+#[macro_use]
 pub mod keyboard;
 pub mod pic_8259;
-pub mod ps2_keyboard;
 pub mod ps2_controller_8042;
 pub mod uart_16550;
 pub mod vga;
